+++ CHANGELOG.md
@@ -4,7 +4,15 @@
 
 ### Snowpark Python API Updates
 
-### Snowpark Local Testing Updates
+#### New Features
+
+- Added support for the following APIs:
+  - snowflake.snowpark.functions
+    - rank
+    - dense_rank
+    - percent_rank
+    - cume_dist
+    - ntile
 
 #### New Features